UPDATE products component
