+++ src/styles/index.css
@@ -1,4 +1,3 @@
-
 /* 
 
     You can write your own styles here...
@@ -10,3 +9,9 @@
     import 'relative/path/to/stylesheet.scss';
 
 */
+
+.cart-fab {
+  position: fixed;
+  bottom: 5vh;
+  width: 100%;
+}
